- unified use of isNan / isFinite across all files that way we only need to modify util.c - added a first pass eliminating warnings issued by gcc 4.7

+++ src/sbml/packages/layout/sbml/test/TestRunner.c
@@ -105,7 +105,7 @@ void
 setTestDataDirectory (void)
 {
   char *srcdir = getenv("srcdir");
-  int  length  = (srcdir == NULL) ? 0 : strlen(srcdir);
+  int  length  = (srcdir == NULL) ? 0 : (int)strlen(srcdir);
   
   
   /**
